Add files via upload
Final Project as an Investment Controller in the Seminar "Finance with Excel" at the University of Zurich.

In this file you will find an Excel Tool I created to conduct the proper calculations for an investment controlling report.

Situation: A pension fund has had a portfolio of several asset classes under active management for roughly 5.5 years. Now the fund's board wants to know what has happened with the portfolio and what adjustments are necessary for the future.

Task: The task of the Investment Controller (me) is to check for legal requirements of the portfolio, annualized returns, risk metrics (Beta, Alpha, Tracking Error and Treynor Ratio), and conduct an profitability analysis of the active management's performance (Performance attribution). The investment policy had estipulated a benchmark at the beginning of the investment period. This benchmark is checked for reliability and the necessary calculations for the comparison are also conducted for the report.

Result: We found out that the portfolio had several flaws in legal requirements, some benchmarks were wrongly chosen, and that the active management had brought more costs than benefits. We recommend proper changes to guarantee future legal requirements in portfolio management for a pension fund and do some recommendations regarding new securities and rebalancing of other positions. Management agreed to every point.